some fixes & extensions

+++ src/shared/array2d.h
@@ -25,7 +25,7 @@ public:
             n >>= 1;
             ++req;
         }
-        if((1 << req) < dim)
+        if((uint32(1) << req) < dim)
             req++;
 
         bool mustcopy = _size || data;
